+++ Sample/SDOSAlamofire/App/Mock/Mock.swift
@@ -7,12 +7,13 @@
 //
 
 import Foundation
+import OHHTTPStubs
 
 let kDefaultsJSONMalformedKey = "kDefaultsJSONMalformedKey"
 let kDefaultsJSONErrorCodeResponseKey = "kDefaultsJSONErrorCodeResponseKey"
 
 func setupServiceMock() throws {
-    OHHTTPStubs.stubRequests(passingTest: { (request: URLRequest) -> Bool in
+    HTTPStubs.stubRequests(passingTest: { (request: URLRequest) -> Bool in
         guard
             let host = request.url?.host,
             let hostToStub = URLComponents(string: Constants.WS.wsUserURL)?.host
@@ -20,7 +21,7 @@ func setupServiceMock() throws {
             return false
         }
         return hostToStub == host
-    }) { _ -> OHHTTPStubsResponse in
+    }) { _ -> HTTPStubsResponse in
         
         let responseData: Data
         if UserDefaults.standard.bool(forKey: kDefaultsJSONMalformedKey) {
@@ -37,7 +38,7 @@ func setupServiceMock() throws {
         }
         
         
-        let response = OHHTTPStubsResponse(data: responseData, statusCode: statusCode, headers: nil)
+        let response = HTTPStubsResponse(data: responseData, statusCode: statusCode, headers: nil)
         response.responseTime = TimeInterval.random(in: 0.3 ... 2)
 
         return response
@@ -45,7 +46,7 @@ func setupServiceMock() throws {
 }
 
 func setupServiceJSONAPIMock() throws {
-    OHHTTPStubs.stubRequests(passingTest: { (request: URLRequest) -> Bool in
+    HTTPStubs.stubRequests(passingTest: { (request: URLRequest) -> Bool in
         guard
             let host = request.url?.host,
             let hostToStub = URLComponents(string: Constants.WS.wsJSONAPIURL)?.host
@@ -53,12 +54,12 @@ func setupServiceJSONAPIMock() throws {
                 return false
         }
         return hostToStub == host
-    }) { _ -> OHHTTPStubsResponse in
+    }) { _ -> HTTPStubsResponse in
         
         let responseData: Data = JSON.correctJSONAPIData
         let statusCode: Int32 = 200
         
-        let response = OHHTTPStubsResponse(data: responseData, statusCode: statusCode, headers: nil)
+        let response = HTTPStubsResponse(data: responseData, statusCode: statusCode, headers: nil)
         response.responseTime = TimeInterval.random(in: 0.3 ... 2)
         
         return response