+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.0 Support for Swift Package Manager and separate JSONAPI](https://github.com/SDOSLabs/SDOSAlamofire/tree/v2.0.0)
+
+- Add support for Swift Package Manager
+- Separate support for JSONAPI in other repo: https://github.com/SDOSLabs/SDOSAlamofireJapx.git
+
 ## [1.0.0 Stable version](https://github.com/SDOSLabs/SDOSAlamofire/tree/v1.0.0)
 
 - Change dependency of Alamofire for use the stable version 5.1.0